+++ pumpkin-lib/src/propagators/cumulative/time_table/mod.rs
@@ -69,10 +69,12 @@
 //! Conference, CP 2015, Cork, Ireland, August 31--September 4, 2015, Proceedings 21, 2015, pp.
 //! 149–157.
 
+mod time_table_over_interval;
 mod time_table_per_point;
 mod time_table_per_point_incremental;
 mod time_table_util;
 
+pub use time_table_over_interval::*;
 pub use time_table_per_point::*;
 pub use time_table_per_point_incremental::*;