+++ pumpkin-lib/src/propagators/cumulative/time_table/time_table_over_interval.rs
@@ -1,8 +1,3 @@
-//! [`Propagator`] for the Cumulative constraint; it
-//! reasons over individual time-points instead of intervals. See [`TimeTablePerPointPropagator`]
-//! for more information.
-
-use std::collections::BTreeMap;
 use std::rc::Rc;
 
 use super::time_table_util::propagate_based_on_timetable;
@@ -10,7 +5,6 @@ use super::time_table_util::should_enqueue;
 use super::time_table_util::ResourceProfile;
 use crate::basic_types::Inconsistency;
 use crate::basic_types::PropagationStatusCP;
-use crate::engine::cp::propagation::ReadDomains;
 use crate::engine::opaque_domain_event::OpaqueDomainEvent;
 use crate::engine::propagation::EnqueueDecision;
 use crate::engine::propagation::LocalId;
@@ -19,6 +13,7 @@ use crate::engine::propagation::PropagationContextMut;
 use crate::engine::propagation::Propagator;
 use crate::engine::propagation::PropagatorConstructor;
 use crate::engine::propagation::PropagatorConstructorContext;
+use crate::engine::propagation::ReadDomains;
 use crate::engine::variables::IntegerVariable;
 use crate::propagators::util::create_inconsistency;
 use crate::propagators::util::create_tasks;
@@ -26,14 +21,29 @@ use crate::propagators::util::reset_bounds_clear_updated;
 use crate::propagators::util::update_bounds_task;
 use crate::propagators::CumulativeConstructor;
 use crate::propagators::CumulativeParameters;
+use crate::propagators::Task;
+#[cfg(doc)]
+use crate::propagators::TimeTablePerPointPropagator;
 use crate::pumpkin_assert_extreme;
+use crate::pumpkin_assert_moderate;
+use crate::pumpkin_assert_simple;
+
+/// An event storing the start and end of mandatory parts used for creating the time-table
+#[derive(Debug)]
+pub struct Event<Var> {
+    /// The time-point at which the [`Event`] took place
+    time_stamp: i32,
+    /// Change in resource usage at [time_stamp][Event::time_stamp], positive if it is the start of
+    /// a mandatory part and negative otherwise
+    change_in_resource_usage: i32,
+    /// The [`Task`] which has caused the event to take place
+    task: Rc<Task<Var>>,
+}
 
 /// [`Propagator`] responsible for using time-table reasoning to propagate the [Cumulative](https://sofdem.github.io/gccat/gccat/Ccumulative.html) constraint
 /// where a time-table is a structure which stores the mandatory resource usage of the tasks at
-/// different time-points - This method creates a resource profile per time point rather than
-/// creating one over an interval (hence the name). Furthermore, the [`TimeTablePerPointPropagator`]
-/// has a generic argument which represents the type of variable used for modelling the start
-/// variables, this will be an implementation of [`IntegerVariable`].
+/// different time-points - This method creates a resource profile over an interval rather than
+/// creating one per time-point (as is done in [`TimeTablePerPointPropagator`]).
 ///
 /// See [Sections 4.2.1, 4.5.2 and 4.6.1-4.6.3 of \[1\]](http://cp2013.a4cp.org/sites/default/files/andreas_schutt_-_improving_scheduling_by_learning.pdf)
 ///  for more information about time-table reasoning.
@@ -41,116 +51,283 @@ use crate::pumpkin_assert_extreme;
 /// \[1\] A. Schutt, Improving scheduling by learning. University of Melbourne, Department of
 /// Computer Science and Software Engineering, 2011.
 #[derive(Debug)]
-pub struct TimeTablePerPointPropagator<Var> {
+pub struct TimeTableOverIntervalPropagator<Var> {
     /// Stores whether the time-table is empty
     is_time_table_empty: bool,
     /// Stores the input parameters to the cumulative constraint
     parameters: CumulativeParameters<Var>,
 }
 
-/// The type of the time-table used by propagators which use time-table reasoning per time-point;
-/// using a [`ResourceProfile`] is more complex than necessary (as [`ResourceProfile::start`] =
-/// [`ResourceProfile::end`]) but it allows for a more unified implementation of methods.
+/// The type of the time-table used by propagators which use time-table reasoning over intervals.
 ///
-/// The key t (representing a time-point) holds the mandatory resource consumption of tasks at
-/// that time (stored in a [`ResourceProfile`]); the [ResourceProfile]s are sorted based on
-/// start time and they are non-overlapping
-pub(crate) type PerPointTimeTableType<Var> = BTreeMap<u32, ResourceProfile<Var>>;
+/// The [ResourceProfile]s are sorted based on start time and they are non-overlapping; each entry
+/// in the [`Vec`] represents the mandatory resource usage across an interval.
+pub(crate) type OverIntervalTimeTableType<Var> = Vec<ResourceProfile<Var>>;
 
-impl<Var> PropagatorConstructor for CumulativeConstructor<Var, TimeTablePerPointPropagator<Var>>
+impl<Var> PropagatorConstructor for CumulativeConstructor<Var, TimeTableOverIntervalPropagator<Var>>
 where
-    Var: IntegerVariable + 'static + std::fmt::Debug,
+    Var: IntegerVariable + 'static,
 {
-    type Propagator = TimeTablePerPointPropagator<Var>;
+    type Propagator = TimeTableOverIntervalPropagator<Var>;
 
     fn create(self, context: PropagatorConstructorContext<'_>) -> Self::Propagator {
         let tasks = create_tasks(&self.tasks, context);
-        TimeTablePerPointPropagator::new(CumulativeParameters::new(tasks, self.capacity))
+        TimeTableOverIntervalPropagator::new(CumulativeParameters::new(tasks, self.capacity))
     }
 }
 
-impl<Var: IntegerVariable + 'static> TimeTablePerPointPropagator<Var> {
-    pub(crate) fn new(parameters: CumulativeParameters<Var>) -> TimeTablePerPointPropagator<Var> {
-        TimeTablePerPointPropagator {
+impl<Var: IntegerVariable + 'static> TimeTableOverIntervalPropagator<Var> {
+    pub fn new(parameters: CumulativeParameters<Var>) -> TimeTableOverIntervalPropagator<Var> {
+        TimeTableOverIntervalPropagator {
             is_time_table_empty: true,
             parameters,
         }
     }
 
-    pub(crate) fn debug_propagate_from_scratch_time_table_point(
+    pub(crate) fn debug_propagate_from_scratch_time_table_interval(
         context: &mut PropagationContextMut,
         parameters: &CumulativeParameters<Var>,
     ) -> PropagationStatusCP {
-        // We first create a time-table per point and return an error if there was
+        // We first create a time-table over interval and return an error if there was
         // an overflow of the resource capacity while building the time-table
-        let time_table = TimeTablePerPointPropagator::create_time_table_per_point_from_scratch(
-            context, parameters,
-        )?;
+        let time_table =
+            TimeTableOverIntervalPropagator::create_time_table_over_interval_from_scratch(
+                context, parameters,
+            )?;
         // Then we check whether propagation can take place
-        propagate_based_on_timetable(context, time_table.values(), parameters)
+        propagate_based_on_timetable(context, time_table.iter(), parameters)
     }
 
-    /// Creates a time-table consisting of [`ResourceProfile`]s which represent rectangles with a
-    /// start and end (both inclusive) consisting of tasks with a cumulative height Assumptions:
-    /// The time-table is sorted based on start time and none of the profiles overlap - generally,
-    /// it is assumed that the calculated [`ResourceProfile`]s are maximal
+    /// Creates a list of all the events (for the starts and ends of mandatory parts) of all the
+    /// tasks defined in `parameters`.
     ///
-    /// The result of this method is either the time-table of type
-    /// [`PerPointTimeTableType`] or the tasks responsible for the
-    /// conflict in the form of an [`Inconsistency`].
-    pub(crate) fn create_time_table_per_point_from_scratch(
+    /// The events are returned in chonological order, if a tie between time points occurs then this
+    /// is resolved by placing the events which signify the ends of mandatory parts first (if the
+    /// tie is between events of the same type then the tie-breaking is done on the id in
+    /// non-decreasing order).
+    fn create_events(
         context: &PropagationContextMut,
         parameters: &CumulativeParameters<Var>,
-    ) -> Result<PerPointTimeTableType<Var>, Inconsistency> {
-        let mut time_table: PerPointTimeTableType<Var> = PerPointTimeTableType::new();
-        // First we go over all tasks and determine their mandatory parts
+    ) -> Vec<Event<Var>> {
+        // First we create a list of events with which we will create the time-table
+        let mut events: Vec<Event<Var>> = Vec::new();
+        // Then we go over every task
         for task in parameters.tasks.iter() {
             let upper_bound = context.upper_bound(&task.start_variable);
             let lower_bound = context.lower_bound(&task.start_variable);
-
             if upper_bound < lower_bound + task.processing_time {
-                // There is a mandatory part
-                for i in upper_bound..(lower_bound + task.processing_time) {
-                    // For every time-point of the mandatory part,
-                    //  add the resource usage of the current task to the ResourceProfile and add it
-                    // to the profile tasks of the resource
-                    let current_profile: &mut ResourceProfile<Var> = time_table
-                        .entry(i as u32)
-                        .or_insert(ResourceProfile::default(i));
-                    current_profile.height += task.resource_usage;
-                    current_profile.profile_tasks.push(Rc::clone(task));
+                // The task has a mandatory part, we need to add the appropriate events to the
+                // events list
 
-                    if current_profile.height > parameters.capacity {
-                        // The addition of the current task to the resource profile has caused an
-                        // overflow
-                        return Err(create_inconsistency(
-                            context,
-                            &current_profile.profile_tasks,
-                        ));
+                // Thus we first add an event for the start of a mandatory part (with positive
+                // resource usage)
+                events.push(Event {
+                    time_stamp: upper_bound,
+                    change_in_resource_usage: task.resource_usage,
+                    task: Rc::clone(task),
+                });
+
+                // Then we create an event for the end of a mandatory part (with negative resource
+                // usage)
+                events.push(Event {
+                    time_stamp: lower_bound + task.processing_time,
+                    change_in_resource_usage: -task.resource_usage,
+                    task: Rc::clone(task),
+                });
+            }
+        }
+        // We will go over the events in chronological order (non-decreasing time_stamp);
+        // this allows us to build the time-table in a single pass
+        events.sort_by(|a, b| {
+            match a.time_stamp.cmp(&b.time_stamp) {
+                // If the time_stamps are equal then we first go through the ends of the mandatory
+                // parts. This allows us to build smaller explanations by ensuring
+                // that we report an error as soon as it can be found
+                std::cmp::Ordering::Equal => {
+                    if a.change_in_resource_usage.signum() != b.change_in_resource_usage.signum() {
+                        // If `a` is the start (end) of a mandatory part and `b` is the end (start)
+                        // of a mandatory part then we need to ensure that
+                        // we go through the end of the mandatory part first
+                        a.change_in_resource_usage.cmp(&b.change_in_resource_usage)
+                    } else {
+                        // If both events are starts or both events are ends then we sort on the
+                        // task id for easier reproducibility
+                        a.task.id.unpack().cmp(&b.task.id.unpack())
+                    }
+                }
+                other_ordering => other_ordering,
+            }
+        });
+
+        events
+    }
+
+    fn check_starting_new_profile_invariants(
+        event: &Event<Var>,
+        current_resource_usage: i32,
+        current_profile_tasks: &[Rc<Task<Var>>],
+    ) -> bool {
+        if event.change_in_resource_usage <= 0 {
+            eprintln!("The resource usage of an event which causes a new profile to be started should never be negative")
+        }
+        if current_resource_usage != 0 {
+            eprintln!("The resource usage should be 0 when a new profile is started")
+        }
+        if !current_profile_tasks.is_empty() {
+            eprintln!("There should be no contributing tasks when a new profile is started")
+        }
+        event.change_in_resource_usage > 0
+            && current_resource_usage == 0
+            && current_profile_tasks.is_empty()
+    }
+
+    /// Creates a time-table based on the provided `events` (which are assumed to be sorted
+    /// chronologically, with tie-breaking performed in such a way that the ends of mandatory parts
+    /// are before the starts of mandatory parts).
+    fn create_time_table_from_events(
+        events: Vec<Event<Var>>,
+        context: &PropagationContextMut,
+        parameters: &CumulativeParameters<Var>,
+    ) -> Result<OverIntervalTimeTableType<Var>, Inconsistency> {
+        pumpkin_assert_extreme!(
+            events.is_empty()
+                || (0..events.len() - 1)
+                    .all(|index| events[index].time_stamp <= events[index + 1].time_stamp),
+            "Events that were passed were not sorted chronologically"
+        );
+        pumpkin_assert_extreme!(
+            events.is_empty()
+                || (0..events.len() - 1).all(|index| events[index].time_stamp
+                    != events[index + 1].time_stamp
+                    || events[index].change_in_resource_usage.signum()
+                        <= events[index + 1].change_in_resource_usage.signum()),
+            "Events were not ordered in such a way that the ends of mandatory parts occurred first"
+        );
+
+        let mut time_table: OverIntervalTimeTableType<Var> = Default::default();
+        // The tasks which are contributing to the current profile under consideration
+        let mut current_profile_tasks: Vec<Rc<Task<Var>>> = Vec::new();
+        // The cumulative resource usage of the tasks which are contributing to the current profile
+        // under consideration
+        let mut current_resource_usage: i32 = 0;
+        // The beginning of the current interval under consideration
+        let mut start_of_interval: i32 = -1;
+
+        // We go over all the events and create the time-table
+        for event in events {
+            if start_of_interval == -1 {
+                // A new profile needs to be started
+                pumpkin_assert_moderate!(
+                    TimeTableOverIntervalPropagator::check_starting_new_profile_invariants(
+                        &event,
+                        current_resource_usage,
+                        &current_profile_tasks
+                    ),
+                    "The invariants for creating a new profile did not hold"
+                );
+
+                // We thus assign the start of the interval to the time_stamp of the event, add its
+                // resource usage and add it to the contributing tasks
+                start_of_interval = event.time_stamp;
+                current_resource_usage = event.change_in_resource_usage;
+                current_profile_tasks.push(event.task);
+            } else {
+                // A profile is currently being created
+
+                // We have first traversed all of the ends of mandatory parts, meaning that any
+                // overflow will persist after processing all events at this time-point
+                if current_resource_usage > parameters.capacity {
+                    // An overflow has occurred due to mandatory parts
+                    return Err(create_inconsistency(context, &current_profile_tasks));
+                }
+
+                // Potentially we need to end the current profile and start a new one due to the
+                // addition/removal of the current task
+                if start_of_interval != event.time_stamp {
+                    // We end the current profile, creating a profile from [start_of_interval,
+                    // time_stamp)
+                    time_table.push(ResourceProfile {
+                        start: start_of_interval,
+                        end: event.time_stamp - 1,
+                        profile_tasks: current_profile_tasks.clone(),
+                        height: current_resource_usage,
+                    });
+                }
+                // Process the current event, note that `change_in_resource_usage` can be negative
+                pumpkin_assert_simple!(
+                    event.change_in_resource_usage > 0
+                        || current_resource_usage >= event.change_in_resource_usage,
+                    "Processing this task would have caused negative resource usage which should not be possible"
+                );
+                current_resource_usage += event.change_in_resource_usage;
+                if current_resource_usage == 0 {
+                    // No tasks have an active mandatory at the current `time_stamp`
+                    // We can thus reset the start of the interval and remove all profile tasks
+                    start_of_interval = -1;
+                    current_profile_tasks.clear();
+                } else {
+                    // There are still tasks which have a mandatory part at the current time-stamp
+                    // We thus need to start a new profile
+                    start_of_interval = event.time_stamp;
+                    if event.change_in_resource_usage < 0 {
+                        // The mandatory part of a task has ended, we should thus remove it from the
+                        // contributing tasks
+                        let _ = current_profile_tasks.remove(
+                            current_profile_tasks
+                                .iter()
+                                .position(|current_task| {
+                                    current_task.id.unpack() == event.task.id.unpack()
+                                })
+                                .expect("Task should have been found in `current_profile`"),
+                        );
+                    } else {
+                        // The mandatory part of a task has started, we should thus add it to the
+                        // set of contributing tasks
+                        pumpkin_assert_extreme!(
+                            !current_profile_tasks.contains(&event.task),
+                            "Task is being added to the profile while it is already part of the contributing tasks"
+                        );
+                        current_profile_tasks.push(event.task);
                     }
                 }
             }
         }
-        pumpkin_assert_extreme!(
-            time_table
-                .values()
-                .all(|profile| profile.start == profile.end),
-            "The TimeTablePerPoint method should only create profiles where `start == end`"
-        );
         Ok(time_table)
+    }
+
+    /// Creates a time-table consisting of [`ResourceProfile`]s which represent rectangles with a
+    /// start and end (both inclusive) consisting of tasks with a cumulative height.
+    ///
+    /// **Assumptions:**
+    /// The time-table is sorted based on start time and none of the profiles overlap - it is
+    /// assumed that the calculated [`ResourceProfile`]s are maximal
+    ///
+    /// The result of this method is either the time-table of type
+    /// [`OverIntervalTimeTableType`] or the tasks responsible for the
+    /// conflict in the form of an [`Inconsistency`].
+    pub(crate) fn create_time_table_over_interval_from_scratch(
+        context: &PropagationContextMut,
+        parameters: &CumulativeParameters<Var>,
+    ) -> Result<OverIntervalTimeTableType<Var>, Inconsistency> {
+        // First we create a list of all the events (i.e. start and ends of mandatory parts)
+        let events = TimeTableOverIntervalPropagator::create_events(context, parameters);
+
+        // Then we create a time-table using these events
+        TimeTableOverIntervalPropagator::create_time_table_from_events(events, context, parameters)
     }
 }
 
-impl<Var: IntegerVariable + 'static> Propagator for TimeTablePerPointPropagator<Var> {
+impl<Var: IntegerVariable + 'static> Propagator for TimeTableOverIntervalPropagator<Var> {
     fn propagate(&mut self, context: &mut PropagationContextMut) -> PropagationStatusCP {
-        let time_table = TimeTablePerPointPropagator::create_time_table_per_point_from_scratch(
-            context,
-            &self.parameters,
-        )?;
+        let time_table =
+            TimeTableOverIntervalPropagator::create_time_table_over_interval_from_scratch(
+                context,
+                &self.parameters,
+            )?;
         self.is_time_table_empty = time_table.is_empty();
         // No error has been found -> Check for updates (i.e. go over all profiles and all tasks and
         // check whether an update can take place)
-        propagate_based_on_timetable(context, time_table.values(), &self.parameters)
+        propagate_based_on_timetable(context, time_table.iter(), &self.parameters)
     }
 
     fn synchronise(&mut self, context: &PropagationContext) {
@@ -180,9 +357,6 @@ impl<Var: IntegerVariable + 'static> Propagator for TimeTablePerPointPropagator<
             context,
             self.is_time_table_empty,
         );
-
-        // Note that the non-incremental proapgator does not make use of `result.updated` since it
-        // propagates from scratch anyways
         update_bounds_task(context, &mut self.parameters.bounds, &updated_task);
         result.decision
     }
@@ -192,18 +366,16 @@ impl<Var: IntegerVariable + 'static> Propagator for TimeTablePerPointPropagator<
     }
 
     fn name(&self) -> &str {
-        "CumulativeTimeTablePerPoint"
+        "CumulativeTimeTableOverInterval"
     }
 
     fn initialise_at_root(&mut self, context: &mut PropagationContextMut) -> PropagationStatusCP {
-        // First we store the bounds in the parameters
         for task in self.parameters.tasks.iter() {
             self.parameters.bounds.push((
                 context.lower_bound(&task.start_variable),
                 context.upper_bound(&task.start_variable),
             ))
         }
-
         self.propagate(context)
     }
 
@@ -211,7 +383,7 @@ impl<Var: IntegerVariable + 'static> Propagator for TimeTablePerPointPropagator<
         &self,
         context: &mut PropagationContextMut,
     ) -> PropagationStatusCP {
-        TimeTablePerPointPropagator::debug_propagate_from_scratch_time_table_point(
+        TimeTableOverIntervalPropagator::debug_propagate_from_scratch_time_table_interval(
             context,
             &self.parameters,
         )
@@ -228,7 +400,7 @@ mod tests {
     use crate::engine::propagation::EnqueueDecision;
     use crate::engine::test_helper::TestSolver;
     use crate::propagators::ArgTask;
-    use crate::propagators::TimeTablePerPoint;
+    use crate::propagators::TimeTableOverInterval;
 
     #[test]
     fn propagator_propagates_from_profile() {
@@ -237,7 +409,7 @@ mod tests {
         let s2 = solver.new_variable(1, 8);
 
         let _ = solver
-            .new_propagator(TimeTablePerPoint::new(
+            .new_propagator(TimeTableOverInterval::new(
                 [
                     ArgTask {
                         start_time: s1,
@@ -267,7 +439,7 @@ mod tests {
         let s1 = solver.new_variable(1, 1);
         let s2 = solver.new_variable(1, 1);
 
-        let result = solver.new_propagator(TimeTablePerPoint::new(
+        let result = solver.new_propagator(TimeTableOverInterval::new(
             [
                 ArgTask {
                     start_time: s1,
@@ -308,7 +480,7 @@ mod tests {
         let s2 = solver.new_variable(0, 6);
 
         let _ = solver
-            .new_propagator(TimeTablePerPoint::new(
+            .new_propagator(TimeTableOverInterval::new(
                 [
                     ArgTask {
                         start_time: s1,
@@ -343,7 +515,7 @@ mod tests {
         let a = solver.new_variable(0, 1);
 
         let _ = solver
-            .new_propagator(TimeTablePerPoint::new(
+            .new_propagator(TimeTableOverInterval::new(
                 [
                     ArgTask {
                         start_time: a,
@@ -391,7 +563,7 @@ mod tests {
         let s2 = solver.new_variable(6, 10);
 
         let mut propagator = solver
-            .new_propagator(TimeTablePerPoint::new(
+            .new_propagator(TimeTableOverInterval::new(
                 [
                     ArgTask {
                         start_time: s1,
@@ -433,8 +605,8 @@ mod tests {
         let s1 = solver.new_variable(6, 6);
         let s2 = solver.new_variable(1, 8);
 
-        let mut propagator = solver
-            .new_propagator(TimeTablePerPoint::new(
+        let _ = solver
+            .new_propagator(TimeTableOverInterval::new(
                 [
                     ArgTask {
                         start_time: s1,
@@ -452,8 +624,6 @@ mod tests {
                 1,
             ))
             .expect("No conflict");
-        let result = solver.propagate_until_fixed_point(&mut propagator);
-        assert!(result.is_ok());
         assert_eq!(solver.lower_bound(s2), 1);
         assert_eq!(solver.upper_bound(s2), 3);
         assert_eq!(solver.lower_bound(s1), 6);
@@ -462,7 +632,7 @@ mod tests {
         let reason = solver.get_reason_int(s2.upper_bound_predicate(3)).clone();
         assert_eq!(
             PropositionalConjunction::from(vec![
-                s2.upper_bound_predicate(6),
+                s2.upper_bound_predicate(9),
                 s1.lower_bound_predicate(6),
                 s1.upper_bound_predicate(6),
             ]),
@@ -481,7 +651,7 @@ mod tests {
         let a = solver.new_variable(0, 1);
 
         let mut propagator = solver
-            .new_propagator(TimeTablePerPoint::new(
+            .new_propagator(TimeTableOverInterval::new(
                 [
                     ArgTask {
                         start_time: a,
@@ -554,7 +724,7 @@ mod tests {
         let a = solver.new_variable(0, 1);
 
         let mut propagator = solver
-            .new_propagator(TimeTablePerPoint::new(
+            .new_propagator(TimeTableOverInterval::new(
                 [
                     ArgTask {
                         start_time: a,
@@ -625,7 +795,7 @@ mod tests {
         let s2 = solver.new_variable(1, 8);
 
         let _ = solver
-            .new_propagator(TimeTablePerPoint::new(
+            .new_propagator(TimeTableOverInterval::new(
                 [
                     ArgTask {
                         start_time: s1,
@@ -651,7 +821,7 @@ mod tests {
         let reason = solver.get_reason_int(s2.lower_bound_predicate(5)).clone();
         assert_eq!(
             PropositionalConjunction::from(vec![
-                s2.lower_bound_predicate(2), /* Note that this not the most general explanation, if s2 could have started at 0 then it would still have overlapped with the current interval */
+                s2.lower_bound_predicate(0),
                 s1.lower_bound_predicate(1),
                 s1.upper_bound_predicate(1),
             ]),
@@ -667,7 +837,7 @@ mod tests {
         let s3 = solver.new_variable(1, 15);
 
         let _ = solver
-            .new_propagator(TimeTablePerPoint::new(
+            .new_propagator(TimeTableOverInterval::new(
                 [
                     ArgTask {
                         start_time: s1,
@@ -702,8 +872,7 @@ mod tests {
             PropositionalConjunction::from(vec![
                 s2.upper_bound_predicate(5),
                 s2.lower_bound_predicate(5),
-                s3.lower_bound_predicate(3), /* Note that s3 would have been able to propagate
-                                              * this bound even if it started at time 0 */
+                s3.lower_bound_predicate(2),
             ]),
             reason
         );